move salt to own file
