+++ framework/contracts/native/account-factory/src/error.rs
@@ -58,7 +58,7 @@ pub enum AccountFactoryError {
     #[error("The caller ({caller}) is not the owner account's manager ({manager}). Only manager can create sub-accounts for its account.", )]
     SubAccountCreatorNotManager { caller: String, manager: String },
 
-    #[error("Expected local account id doesn't match, expected: {predicted}, actual: {actual}". Try again.)]
+    #[error("Expected local account id doesn't match, expected: {predicted}, actual: {actual} Try again.")]
     ExpectedAccountIdFailed {
         predicted: AccountId,
         actual: AccountId,
