+++ framework/packages/abstract-core/src/objects/mod.rs
@@ -10,6 +10,7 @@ mod entry;
 pub mod nested_admin;
 pub mod oracle;
 pub mod pool;
+pub mod salt;
 
 pub use pool::*;
